style(feeds): F3b-4 clear golangci/staticcheck/CodeQL findings
- authority_test: bytes.Equal over string comparison (gocritic stringXbytes)
- classifyAcquireError: name the (class, httpStatus) results (gocritic unnamedResult)
- backoffDelay: signed shift count, drop the uint() conversion (gosec G115)
- noteActivation: drop the always-200 httpStatus param (unparam), hardcode 200
- noteOverrides: wire it (was unused) so override_count/revision are published
  after each activation/recovery via runtime.noteOverrideFootprint
- split TestF3b4_Status_DerivedState* to stay under the cognitive-complexity bound
- sanitizeLog the lifecycle-armed log's enum strings (CodeQL CWE-117)

+++ saas_feed_scheduler.go
@@ -191,7 +191,9 @@ func backoffDelay(failures int) time.Duration {
 	if shift > saasFeedBackoffShift {
 		shift = saasFeedBackoffShift
 	}
-	d := saasFeedBackoffMin << uint(shift)
+	// shift is clamped to [0, saasFeedBackoffShift]; a signed shift count is valid Go
+	// (>=1.13) and non-negative here, so no unsigned conversion is needed (avoids G115).
+	d := saasFeedBackoffMin << shift
 	if d > saasFeedBackoffMax || d <= 0 {
 		return saasFeedBackoffMax
 	}
